Proyecto/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
aiofiles==23.2.1
pillow==10.3.0
pydantic[email]==1.10.21
email-validator==2.1.1
